// src/types/issue.ts
import { GithubLabel } from "./label";

export type GithubIssue = {
    url: string;
    repository_url: string;
    labels_url: string;
    comments_url: string;
    events_url: string;
    html_url: string;
    id: number;
    node_id: string;
    number: number;
    title: string;
    user: User;
    labels: GithubLabel[];
    state: State;
    locked: boolean;
    assignee: null;
    assignees: unknown[];
    milestone: null;
    comments: number;
    created_at: string;
    updated_at: Date;
    closed_at: null;
    author_association: AuthorAssociation;
    type: null;
    active_lock_reason: null;
    sub_issues_summary?: SubIssuesSummary;
    issue_dependencies_summary?: IssueDependenciesSummary;
    body: string;
    closed_by: null;
    reactions: Reactions;
    timeline_url: string;
    performed_via_github_app: null;
    state_reason: null;
    draft?: boolean;
    pull_request?: PullRequest;
}

export enum AuthorAssociation {
    Collaborator = "COLLABORATOR",
    Contributor = "CONTRIBUTOR",
    Member = "MEMBER",
    None = "NONE",
}

export interface IssueDependenciesSummary {
    blocked_by: number;
    total_blocked_by: number;
    blocking: number;
    total_blocking: number;
}

export interface PullRequest {
    url: string;
    html_url: string;
    diff_url: string;
    patch_url: string;
    merged_at: null;
}

export interface Reactions {
    url: string;
    total_count: number;
    "+1": number;
    "-1": number;
    laugh: number;
    hooray: number;
    confused: number;
    heart: number;
    rocket: number;
    eyes: number;
}

export enum State {
    Open = "open",
    Close = "close"
}

export interface SubIssuesSummary {
    total: number;
    completed: number;
    percent_completed: number;
}

export interface User {
    login: string;
    id: number;
    node_id: string;
    avatar_url: string;
    gravatar_id: string;
    url: string;
    html_url: string;
    followers_url: string;
    following_url: string;
    gists_url: string;
    starred_url: string;
    subscriptions_url: string;
    organizations_url: string;
    repos_url: string;
    events_url: string;
    received_events_url: string;
    type: Type;
    user_view_type: UserViewType;
    site_admin: boolean;
}

export enum Type {
    User = "User",
}

export enum UserViewType {
    Public = "public",
}